backend/pyproject.toml
[tool.poetry]
name = "nest-adversarial-sandbox"
version = "1.0.0"
description = "AI Safety Testing Platform - Backend API"
authors = ["NEST Team"]
readme = "../README.md"
packages = [{include = "services"}, {include = "scenarios"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"

# AI Provider SDKs
google-generativeai = "^0.8.3"
openai = "^1.54.0"
anthropic = "^0.52.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]
